Introduce new resource type for latency control
In this patchset a new resource type is introduced, it is intended to
control network latency by request. Right now only explicit ReportData
calls are handled, but it is possible to bind it with some system event
later. Concrete latency switchers could be implemented, but right now
only linux PMQoS is done.

Signed-off-by: Petrov Maxim <petrov.maxim@huawei.com>